# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your@mail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/lti-form/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete lti integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTIForm"
                ],
                "summary": "delete lti integration",
                "parameters": [
                    {
                        "description": "lti id",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormDeleteInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormDeleteResponse"
                        }
                    }
                }
            }
        },
        "/v1/lti-form/get": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get lti integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTIForm"
                ],
                "summary": "get lti integration",
                "parameters": [
                    {
                        "description": "lti id",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormGetInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormGetResponse"
                        }
                    }
                }
            }
        },
        "/v1/lti-form/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List lti integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTIForm"
                ],
                "summary": "list lti integration",
                "parameters": [
                    {
                        "description": "lti list info",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormListInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormEditResponse"
                        }
                    }
                }
            }
        },
        "/v1/lti-form/upsert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a lti integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTIForm"
                ],
                "summary": "create lti integration",
                "parameters": [
                    {
                        "description": "lti form info",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormEditInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.LtiFormEditResponse"
                        }
                    }
                }
            }
        },
        "/v1/pnet-server/delete": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete pnet-server integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PNETServer"
                ],
                "summary": "delete pnet-server integration",
                "parameters": [
                    {
                        "description": "pnet-server id",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerDeleteInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerDeleteResponse"
                        }
                    }
                }
            }
        },
        "/v1/pnet-server/get": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get lti integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PNETServer"
                ],
                "summary": "get lti integration",
                "parameters": [
                    {
                        "description": "pnet-server id",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerGetInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerGetResponse"
                        }
                    }
                }
            }
        },
        "/v1/pnet-server/list": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List pnet server integration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PNETServer"
                ],
                "summary": "list pnet server integration",
                "parameters": [
                    {
                        "description": "pnet-server list info",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerListInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerEditResponse"
                        }
                    }
                }
            }
        },
        "/v1/pnet-server/upsert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a pnet server connection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PNETServer"
                ],
                "summary": "create PNETServer integration",
                "parameters": [
                    {
                        "description": "pnet server info",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerEditInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/usecases.PNETServerEditResponse"
                        }
                    }
                }
            }
        },
        "/v1/token/renew": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Renew access and refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "summary": "renew access and refresh tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/lti": {
            "get": {
                "description": "Базовая информация для интеграции LTI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTI"
                ],
                "summary": "получить базовую информацию для интеграции",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecases.LtiFormEditOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/v2/lti/launch/": {
            "post": {
                "description": "инициализация подключения к LTI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTI"
                ],
                "summary": "инициализация подключения к LTI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecases.LtiFormEditOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/v2/lti/launch/authenticate/": {
            "post": {
                "description": "аутентификация LTI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTI"
                ],
                "summary": "аутентификация через LTI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecases.LtiFormEditOutputDTO"
                            }
                        }
                    }
                }
            }
        },
        "/v2/lti/launch/public_keys/": {
            "get": {
                "description": "RSA ключи для интеграции LTI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LTI"
                ],
                "summary": "получить RSA ключи для интеграции LTI",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/usecases.LtiFormEditOutputDTO"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LTIForm": {
            "type": "object",
            "required": [
                "created_at",
                "lti_key",
                "lti_secret",
                "name",
                "updated_at",
                "version"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lti_key": {
                    "type": "string"
                },
                "lti_secret": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.PNETServer": {
            "type": "object",
            "required": [
                "created_at",
                "updated_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_online_status": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "unit_rate": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "usecases.LtiFormDeleteInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.LtiFormDeleteResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.LtiFormDeleteInputDTO"
                }
            }
        },
        "usecases.LtiFormEditInputDTO": {
            "type": "object",
            "required": [
                "name",
                "version"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "usecases.LtiFormEditOutputDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.LtiFormEditResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.LtiFormEditOutputDTO"
                }
            }
        },
        "usecases.LtiFormGetInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.LtiFormGetOutputDTO": {
            "type": "object",
            "properties": {
                "model": {
                    "$ref": "#/definitions/models.LTIForm"
                }
            }
        },
        "usecases.LtiFormGetResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.LtiFormGetOutputDTO"
                }
            }
        },
        "usecases.LtiFormListInputDTO": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
        "usecases.PNETServerDeleteInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.PNETServerDeleteResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.PNETServerDeleteInputDTO"
                }
            }
        },
        "usecases.PNETServerEditInputDTO": {
            "type": "object",
            "required": [
                "name",
                "token",
                "url"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "unit_rate": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "usecases.PNETServerEditOutputDTO": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.PNETServerEditResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.PNETServerEditOutputDTO"
                }
            }
        },
        "usecases.PNETServerGetInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "usecases.PNETServerGetOutputDTO": {
            "type": "object",
            "properties": {
                "model": {
                    "$ref": "#/definitions/models.PNETServer"
                }
            }
        },
        "usecases.PNETServerGetResponse": {
            "type": "object",
            "required": [
                "error",
                "msg"
            ],
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "msg": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/usecases.PNETServerGetOutputDTO"
                }
            }
        },
        "usecases.PNETServerListInputDTO": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}